Many changes
1, dataTypes in json expression format, convert to TapMapping and TapType, Use target TapMapping and TapType to generate the best data types.
2, Generate the best data types is not finished, very simple implementation. Need work it out later.
3, Prepared TDD Connector source input. later will use it as a test connector.
4, TargetNodeDriver invoke createTables method with the generated data types.

+++ tapdata-api/src/main/java/io/tapdata/entity/codec/impl/TapNumberCodec.java
@@ -17,7 +17,7 @@ public class TapNumberCodec implements ToTapValueCodec<TapNumberValue>, FromTapV
             tapNumber = (TapNumber) tapType;
         }
         if(tapNumber != null) {
-            Long scale = tapNumber.getScale();
+            Integer scale = tapNumber.getScale();
             //TODO need more code
             if(scale == null || scale == 0)
                 return tapValue.getValue().longValue();
